ozmaps: mapmanager: remove %llu compiler error
compiler error = warning: format ‘%llu’ expects argument of type
 ‘long long unsigned int’, but argument 2 has type ‘uint64_t
resolved by typecasting printf variable to (long long) before printing
Since long long has different meanings on 32 bit vs 64 bit machines
 typecasting is easiest solution for portability.

+++ src/ozmaps/mapmanager.c
@@ -103,7 +103,8 @@ SDL_Surface *getmap(int mapflags,      /* instructions on which map to get */
       if ((mapindexTemp = findgpsmap(gps, mapsetNow)) == NULL) { // not found
         mapset++;
         if (bookend == 9999) {
-          printf(" no map found for gps (dddmmssddmmss)=%llu \n", gps);
+          printf(" no map found for gps (dddmmssddmmss)=%llu \n",
+                 (long long)gps);
           exit(0);
         }
       } else { // found
